Implement HandleLabel
